Reorganize memory and state into one NesState struct
I do NOT like the code changes that resulted in cpu.rs, especially with regards to all the extra temporary values needed to satisfy the borrow checker.

Not sure how to resolve; Rust's borrow checker is due for improvement that might fix it. but I also feel like the entire NesState might not be needed here if memory could be implemented in a way that didn't need the rest of the system for its side-effects.