Implement AxROM; Battletoads boots! Doesn't exactly... play, but it boots!
I didn't expect Battletoads to do much of anything resembling gameplay,
but I'm pleasantly surprised that it gets as far as it does. It's going
to be an uphill battle that one, for starters the CPU needs to spend
real cycles on its instructions, and stop pretending that everything is
1-cycle long.

Mostly though, I'm really happy that it was so easy to add in the new
mapper; that system works really well so far.