Begin work on some utility font and drawing functions
This is half-baked and unfinished, but I want to go through with it
eventually. The built in text drawing routines in Piston are okay,
but a little too complex for what I need to do, and I want to have
pixel-level control over the final layout. The plan is to implement
software based bitmap fonts instead.