Implement the skeleton of a newer, more cycle accurate CPU
This presently does nothing but pass instructions through to the old
implementation. A side-effect of this behavior is that instructions
now take a minimum of 2 CPU cycles to complete, which changes many
timing-related effects in games, though the wrongness is merely
different in these cases. No real 6502 instruction completes in
less than 2 CPU cycles, so this behavior is more correct than before.