Correct PPU read address during invalid sprite fetches
When reading tile data for inactive sprites, we ended up adding a
Y coordinate value that is greater than 16 quite often, which
causes the read address to fly off the deep end. This is mostly
harmless, except for addresses that fall outside of the 0x1FFF
range, which fail to properly read from the mapper. This was
breaking scanline counting in the MMC3 mapper, and would have
caused other issues.

I'm not sure what the correct fix is here, as the presice behavior
of these reads seems to be poorly defined; the tile index is set
to 0xFF, but what about the Y offset? Where is that calculated
exactly? I'd like to do some research and nail this behavior down.
For now, we fix this by forcing the tile to at least be within the
selected pattern table, which performs the correct read behavior
from MMC3's point of view, and causes A12 to consistently oscillate
correctly, even for inactive sprites, and even in 8x16 mode.