Add debug_read_byte to Mapper, implement for MMC3
This is a special read_byte which is meant to be passive, the mapper
should NOT update its internal state, and should instead passively
return whatever would be returned for such a read. This should be
called by any debugging routines that perform reads outside of what
the NES program would do, to avoid corrupting the game's expected
behavior when reads have side effects.