Implement basic sanity checks for incoming cartridge data
Notably, the magic header is now validated, and attempting to load
a ROM with an unsupported mapper type will bail before completing
the load. Before it would load a dummy mapper and then usually crash
the program with out of bounds reads immediately, so this is an
improvement. Always best to die on purpose.