Fix a bug with APU's Pulse 1 channel in sweep negate mode
When the sweep unit is in negate mode, it's meant to subtract the change
amount from the current period to determine the target period. When the
sweep amount is 0, on the Pulse 1 channel, this calculation results in
the value -1. That would overflow the u16 in-memory representation of
the target period (whoops!) so we add a special case to avoid the
overflow, and return 0 instead. I'm not entirely sure what real hardware is
supposed to do in this case, but this appears to fix compatability with
games that disable the sweep channel in this manner, including the Duck
Tales series.