Implement passive PPU reads and heatmap, similar to CPU memory
The heatmap is mostly useful for debugging purposes, and simply tracks the last 20 or so reads in order. Shells can read this data to highlight recently accessed memory.

Passive PPU reading is equally useful for debugging purposes, and I believe this corrects a minor bug when doing so from the IO register view.