Finish implementing all missing unofficial instructions
It's taken surprisingly long to reach this point, but the emulator now has 100% coverage of all possible opcodes. It now passes a great deal more CPU tests, including those that specifically test undefined operation. I'm far from done with the CPU, but this is a big step in the right direction.