Implement open bus behavior, including in mappers
I chose to have mappers return an Option<u8> as that seems the most sensible from a Rust code standpoint. A value of None means that no circuitry is connected at the requested address. Both the CPU and PPU will rely on Open Bus behavior in these cases, documented here:

https://wiki.nesdev.com/w/index.php/Open_bus_behavior

With this commit. the cpu_exec_space test now passes.