Actually implement PxRom CHR switching as a latch
Before we would simply copy the values over when detecting the special
tiles, but this neglected the very important case of writing the desired
banks and never sending any special tiles. If you write to the active
bank, it needs to update right away. Implementing the latch corrects
this behavior, and finally Punch-Out behaves correctly and seems to have
no more rendering bugs.