Begin moving all mirroring logic into the Mapper
This is just going to make more sense, with how many strange edge
cases exist. Strictly the VRAM is supposed to live inside the
NES console, but in practice it's much simpler to model the
"NES VRAM" as being something the mapper owns and controls. This
is logically equivalent, and enables the complex logic needed to
support advanced mappers without trying to stuff all the
possible configurations into a giant list of enums.