Palette reads from the PPU should not end up on open_bus
The PPU's palette memory shadows the bus, but is a separate thing.
Memory access from within the PPU should not pass through to the
mapper, and memory access from the CPU has an odd interaction with
the read buffer. This change approximates the behavior as best
as I currently understand it. More testing might be needed.