Correct an off-by-one error in the DPCM APU timing table
The table itself is correct, but when reloading the timer, we need
to account for one extra clock. Before this change, though it is
hard to hear in most retail games, the DPCM channel was ever so
slightly flat. This probably affects APU timing tests that check
the DPCM status flags.