Invert triangle output, based on results from hardware testing
Although the Triangle appears to start at 15 when inspecting
Visual 6502, it appears the output is somehow inverted on its
journey through the mixer. The simplest way to model its observed
behavior is to begin the triangle sequence counter at 0, instead
of 15, which produces results more consistent with real hardware.
First observed by Kitrinx in #emulator-development, on their
Famicom AV, verified on my frontloading NTSC NES.

This information is still under research, as it uncovers a series
of assumed errors on the nesdev wiki that now need exploration.