For now, entirely skip the NES's high and low pass filter chain
The result of applying these filters is incredibly muddy, which
is almost certainly a flaw of the technique. I suspect it may need
a much higher source sample rate, but for the moment am simply
disabling the effect. I'd rather have clean, slightly wrongly
mixed audio than a muffled mess.