Introduce abstractions for audio channel semantics
Basically, for any given channel, I want to be able to report simple
frequency, volume, and patch data, for use with visualizers. These
are not meant to be fully accurate with respect to all of the
raw channel data (for that, specilize on the Mapper and read the
registers directly) but more useful for high level renderings.