Add facility to store and optionally dump raw, 1.7 MHz audio
Not that Audacity has the ability to *process* audio at this
high of a sample rate without workarounds, but one can't have
everything. This is useful mostly because I am bad at audio
filters, and bypassing the high and low pass results in a bit
of a cleaner signal for hardware comparisons.