More formally define the FamiCom and Nes filter chain
Technically this allows switching between the two in the consuming
application, though none of the reference implementations support
this quite yet. Filter design pulled from here:

https://wiki.nesdev.com/w/index.php/APU_Mixer

With a properly implemented IIR, this doesn't sound hideous in
NES mode. However, as iNes headers provide no way to disambiguate
NES and FamiCom, and there is thus no reliable signal for which
filter set should be used, the program will always default in the
minimal FamiCom set, as it is especially more accurate when used
alongside expansion audio.