Implement basic support for Mapper 26
This mapper is also VRC6, with one differnce: the address lines
A0 and A1 are swapped, which means $xxx0 becomes $0001 and vice
versa. To implement this, we detect the incoming mapper number,
manually swap the lines, then decode normally. This seems to work
well for both games using the mapper, one of which is still
glitchy for unrelated CHR ROM mirroring reasons.