*Properly* compute the VRC6's mixing level
Previously we sorta guessed at a 0.5 multiplier, but as it turns
out, we want the contributiuon from one full-volume VRC6 pulse
channel to match the contribution of one full-volume 2A03 pulse
channel. We have that math on hand, so we compute this properly,
ends up being around 0.6 or so, but I'm keeping the full equasions
in here rather than committing a magic numbner.