Replace the 203 entry approximation with the full formula
This still uses a LUT, but this time it is a *proper* LUT with
all 65536 values. Still a negligible amount of data on all target
platforms, and just as quick to index.

The audible result is subtle; basically the noise and DPCM channels
are a smidge louder in the mix.