Set the audio buffer size to the smallest usable power of 2
The default of 4096 has been fine up to this point, but it does
create a noticeable lag behind the visuals onscreen. Instead,
we'd like to prefer about one full frame's worth of buffer for
all practical applications, and since the sample rate is not
set in stone, we compute this new requirement on the fly.

The power of 2 thing is simply because a good deal of audio
hardware will expect buffers of this size anyway.