For NSF only, initialize VRC6 before handing off to the ROM
Some older NSF files appear to expect basic initialization of the
expansion chips to be performed. Relying on this is out of spec,
but in the interest of compatibility, let's assume that any NSF
which indicates it will use an expansion audio chip is definitely
planning to actually use it, and un-mute those chips in advance.

(As I find good test cases, this will probably need to be done for
 the other expansions, and we might want to re-initialize them
 when switching tracks)