Place the contents of 'v' on the PPU address bus when rendering ends
This is done with a dummy read due to the mapper interface. Mappers
which care about the value on the bus can monitor PPU reads and
writes, and track the address over time this way. That only works
if all changes make it to one of those functions in one form or
another. This particular change is needed to emulate MMC3 quirks
with sprites in $0000, affecting an invisible scanline glitch in
Wario's Woods, and a visible scanline glitch in several homebrew
titles.