Clock MMC3 IRQ based on filtered A12
The basic strategy here is supposed to be for filtered A12 to get
set on rising edges, but cleared on the third M2 clock in a row
while real A12 is low. In practice, for yet unclear reasons, four
M2 clocks are needed instead of 3. The community is still
investigating the true cause, so for the moment we're having
RusticNES emulate this as 4x M2 clocks to maintain compatability.