Implement N163 audio
Right now this is doing the hardware-accurate "hold one channel
at a time" thing, and depending on the number of enabled channels
this produces an audible, high-pitched whine. Real hardware is
normally followed by an RF modulator and other sources of low pass
that would attenuate this signal; right now RusticNES is set up to
try very hard not to do that, so the whine is probably more audible
than it would be on real hardware. I don't actually have a real
FamiCom to compare with, so we'll just have to deal with it by
ear.