Greatly improve MMC3 accuracy, based on the somewhat recent decap
Most of the actual work here was tracking down various places in the
PPU where the address lines weren't behaving correctly from the
mapper's point of view. Turns out MMC3 is really sensitive to these
sorts of things when its M2 cooldown is dialed in to what the
hardware expects.