Let's run with the MIT license for the core project
Note: Please credit authorship as zeta0134, as this name is better
known throughout the communities that inspired the work than my
legal name. To clarify this author's intent: I feel that if you are
bundling rusticnes-core as the engine driving a standalone NES game,
simply placing attribution somewhere in the install folder is sufficient.
There is no need to modify the NES file or display anything at runtime,
save those bytes for something more useful.