Tune filtering parameters to remove aliasing artifacts
Turns out having the initial IIR be just half the intermediate
sample rate is not nearly enough, and was leaving behind a lot of
junk. So, instead of a 0.5 cutoff we switch to a 0.25 cutoff and
that clears up most of the aliasing issues.

While doing this, I noticed that N163 games in particular were
gaining unusual harmonics, and even though N163 is full of artifacts
in genreal these were weird. Turns out having the intermediate
sample rate be an *exact* multiple of the final sample rate was
introducing some kind of phasing, I'm guessing due to floating
point rounding errors in the various accumulating clocks. We fix
that by adding pi/32 to the intermediate samplerate, which gives
it an irrational decimal portion and eliminates the possibility of
it being an even multiple of the target sample rate, no matter what
that sample rate is.

We need to expand the window_sinc filter a little bit to compensate
for the larger intermediate sample rate, but otherwise it sounds
pretty good. This is still a big improvement over the original
160-entry window sinc and doing all of the filtering at 1.7 MHz.