Pass the NMI and IRQ vectors from the NSF file into memory
Note that any NSF attempting to *use* these has the potential to
break the player, but on principal we shouldn't hide any memory
that we don't need to repurpose for ourselves. Right now the NSF
player only needs the reset vector, so that is all it will override.

(If we wanted to be especially fancy, we could have that override
 expire after being read the first time.)